* commit '984f443541b271876cc9087194cbd5a0e3b26326':
  TM-14872 - Added confirmation popup to the Notices

+++ web-app/tds/web-app/app-js/modules/noticeManager/components/view-edit/notice-view-edit.component.ts
@@ -65,10 +65,15 @@ export class NoticeViewEditComponent {
 	}
 
 	protected deleteNotice(): void {
-		this.noticeService.deleteNotice(this.model.id.toString())
-			.subscribe(
-				res => this.activeDialog.close(),
-				error => this.activeDialog.dismiss(error));
+		this.promptService.open('Confirmation Required', 'You are about to delete the selected notice. Do you want to proceed?', 'Yes', 'No')
+			.then((res) => {
+				if (res) {
+					this.noticeService.deleteNotice(this.model.id.toString())
+						.subscribe(
+							res => this.activeDialog.close(),
+							error => this.activeDialog.dismiss(error));
+				}
+			});
 	}
 
 	/**
